Let Vespa have its own worker and queue.
Every search database behaves a bit differently. Vespa takes quite long to index documents (as those indexes are directly available). So we make sure here that each full text search database can be decide for itself that if it need some background task for indexing or do it directly in the request itself.