Fix warning about post generation calls of factory boy
Factory boy won't save the instance in post generation hooks in the next major version. I fixed this here, but the warning still remains (I guess more as a hint). So we still ignore the warning and remove this ignore when the next major version of factory boy gets released.